further split folder structure for embedded tests, add embedded2Join sample schemata
